:construction: Working on multiple tile colisions

+++ Engine/Vector.py
@@ -3,14 +3,9 @@ from typing import Tuple
 
 
 class Vector():
-    def __init__(self, x: float = 0, y: float = 0) -> None:
+    def __init__(self, x: int = 0, y: int = 0) -> None:
         self.x = x
         self.y = y
-    
-    # def add(self, vector) -> None:
-    #     self.x += vector.x
-    #     self.y += vector.y
-    #     return self
 
     def get(self) -> Tuple:
         return (self.x, self.y)
